Add int cast to suppress gcc warning

+++ Window.hpp
@@ -239,11 +239,11 @@ public:
             if (s.poses.has_value()) {
                 if (key == GLFW_KEY_J) {
                     s.pose_idx -= 1;
-                    if (s.pose_idx < 0) s.pose_idx += s.poses->shape[0];
+                    if (s.pose_idx < 0) s.pose_idx += (int)(s.poses->shape[0]);
                 }
                 if (key == GLFW_KEY_K) {
                     s.pose_idx += 1;
-                    s.pose_idx %= s.poses->shape[0];
+                    s.pose_idx %= (int)(s.poses->shape[0]);
                 }
             }
         }
